requirements/base.txt
Django==2.1.7
django-cleanup==3.2.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.2
moviepy==1.0.0
psycopg2==2.7.7
